+++ src/Setup.ts
@@ -5,3 +5,8 @@ const moment = require('moment');
 Date.prototype.toJSON = function () {
   return moment(this).format();
 };
+
+import Translations from './Utilities/Format/Translations';
+
+import messages from './Translations.json';
+Translations.addMessages(messages);