+++ src/Utilities/Format/Text.ts
@@ -1,0 +1,11 @@
+class Text {
+  static beginsWith = (text: string, ...prefixes: string[] ): boolean => (
+    prefixes.some(prefix => text.substring(0, prefix.length) === prefix)
+  )
+
+  static endsWith = (text: string, ...suffixes: string[]) => (
+    suffixes.some(suffix => text.substring(text.length -suffix.length) === suffix)
+  )
+}
+
+export default Text;