+++ src/Utilities/Notifier.ts
@@ -1,0 +1,7 @@
+class Notifier {
+  static warn = (message: string) => console.warn(
+    `WARNING: ${message}`
+  );
+}
+
+export default Notifier;