+++ src/Utilities/index.ts
@@ -1,0 +1,11 @@
+import Provider from './Provider';
+import Notifier from './Notifier';
+import Format from './Format';
+
+class Utilities {
+  static Provider = Provider;
+  static Notifier = Notifier;
+  static Format = Format;
+}
+
+export default Utilities;